Rocket league update

+++ include/Mesh.hpp
@@ -34,7 +34,7 @@ private:
 
     void updateBuffers();
 
-    std::vector<std::uint16_t> indices;
+    std::vector<std::uint32_t> indices;
     std::vector<glm::vec3> vertices;
     std::vector<glm::vec3> normals;
     std::vector<glm::vec2> texCoords;
